Make fonts integers, add font loading

+++ Ikala-Graphics/src/main/java/com/ikalagaming/graphics/frontend/gui/data/DrawList.java
@@ -973,13 +973,13 @@ public class DrawList {
 
     }
 
-    public void addText(Font Font, float fontSize, float posX, float posY, int color, String text) {
+    public void addText(Font Font, int fontSize, float posX, float posY, int color, String text) {
         // TODO(ches) implement this
     }
 
     public void addText(
             Font Font,
-            float fontSize,
+            int fontSize,
             float posX,
             float posY,
             int color,
@@ -990,7 +990,7 @@ public class DrawList {
 
     public void addText(
             Font Font,
-            float fontSize,
+            int fontSize,
             float posX,
             float posY,
             int color,
