+++ Ikala-Graphics/src/main/java/com/ikalagaming/graphics/frontend/gui/data/FontStackInfo.java
@@ -2,6 +2,6 @@ package com.ikalagaming.graphics.frontend.gui.data;
 
 public class FontStackInfo {
     private Font font;
-    private float sizeBeforeScaling;
-    private float sizeAfterScaling;
+    private int sizeBeforeScaling;
+    private int sizeAfterScaling;
 }
